requirements.txt
pandas>=2.0.0
plotly>=5.18.0
nltk>=3.8.0
scikit-learn>=1.3.0
pyobjc-framework-Contacts>=10.0
pyarrow>=14.0.0
emoji>=2.8.0
